Make REPL shared
